refactor[BE]: Fix login api url
