+++ backend/src/api/auth/auth.module.ts
@@ -1,13 +1,23 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { PrismaModule } from 'src/prisma/prisma.module';
-import { JwtService } from '@nestjs/jwt';
+import { AuthController } from './auth.controller';
+import { JwtModule } from '@nestjs/jwt';
+import { PassportModule } from '@nestjs/passport';
 
 @Module({
 	imports: [
 		PrismaModule,
+		PassportModule,
+		JwtModule.registerAsync({
+			useFactory: () => ({
+			  secret: process.env.JWT_SECRET,
+			  signOptions: { expiresIn: 3600 },
+			})
+		}),
 	],
-	providers: [AuthService, JwtService],
+	providers: [AuthService],
+	controllers: [AuthController],
 	exports: [AuthService],
 })
 export class AuthModule {}
