+++ backend/src/api/users/users.module.ts
@@ -3,25 +3,16 @@ import { UsersService } from './users.service';
 import { PrismaModule } from '../../prisma/prisma.module';
 import { UsersController } from './users.controller';
 
-import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
-import { AuthService } from '../auth/auth.service';
 import { FTStrategy } from '../auth/ft.strategy';
 
 @Module({
   imports: [
     PrismaModule,
     PassportModule, 
-    JwtModule.registerAsync({
-        useFactory: () => ({
-          secret: process.env.JWT_SECRET,
-          signOptions: { expiresIn: 3600 },
-        })
-		  }),
   ],
   providers: [
     UsersService, 
-    AuthService, 
     FTStrategy, 
   ],
   controllers: [
