+++ backend/src/app.module.ts
@@ -3,6 +3,7 @@ import { ConfigModule } from '@nestjs/config';
 import { UsersModule } from './api/users/users.module';
 import { PrismaModule } from './prisma/prisma.module';
 import { MyModule } from './api/my/my.module';
+import { AuthModule } from './api/auth/auth.module';
 
 @Module({
   imports: [
@@ -12,6 +13,7 @@ import { MyModule } from './api/my/my.module';
     UsersModule, 
     MyModule,
     PrismaModule,
+    AuthModule,
   ],
   controllers: [],
   providers: [],
